// client/src/enums/gameCardEnums.ts
export enum CardValues {
    seven = '7',
    eight = '8',
    nine = '9',
    jack = 'J',
    queen = 'Q',
    king = 'K',
    ten = '10',
    ace = 'A',
}

export enum CardSuits {
    spades = 'spades',
    diamonds = 'diamonds',
    clubs = 'clubs',
    hearts = 'hearts',
}

export enum Deals {
    all = 'all',
    one = 'one',
    pants = 'pants',
    satellite = 'satellite',
}